fix: normalize replication ports
